Updates template.
To avoid corruption, replace watermark elements with something valid:

e.g.,
<< /Type /Font /Subtype /Type1 /BaseFont /Helvetica /Encoding
/WinAnsiEncoding >>

The obj numbers are sequential and cannot be left our or the pdftk
parser on Heroku will have problems.

Search for 'Water', replace the contents of the related object along
with the next 3 obj nodes (hyperlinks and box with black border --
button).